Migrate approval to use bigint

+++ tg_bot/modules/sql/approve_sql.py
@@ -1,6 +1,7 @@
 import threading
 
-from sqlalchemy import Column, String, UnicodeText, Integer, func, distinct
+from sqlalchemy import Column, String
+from sqlalchemy.sql.sqltypes import BigInteger
 
 from tg_bot.modules.sql import BASE, SESSION
 
@@ -8,7 +9,7 @@ from tg_bot.modules.sql import BASE, SESSION
 class Approvals(BASE):
     __tablename__ = "approval"
     chat_id = Column(String(14), primary_key=True)
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
 
     def __init__(self, chat_id, user_id):
         self.chat_id = str(chat_id)  # ensure string
